Move input components to components/inputs
as per cleanup directions in #308.
